refactor: extract MCP interface package
